requirements.txt
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
imageio==2.10.1
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
networkx==2.6.3
numba==0.54.1
numpy==1.21.3
opencv-python==4.5.4.58
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
Pygments==2.10.0
pyparsing==3.0.4
python-dateutil==2.8.2
PyWavelets==1.1.1
pywin32==302
pyzmq==22.3.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.10.12
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
